Pull out deleteCoffe into it's own component, start stripping out fetchTrigger
