+++ my-app/src/pages/EquipmentPage.js
@@ -12,28 +12,26 @@ function EquipmentPage({
   onAddToCart,
 }) {
   return (
-    <>
-      <div className="equipment-page">
-        <div className="container">
-          <AddNewEquipment onAddEquipment={onAddEquipment} />
-          <Container fluid>
-            <Row>
-              {equipmentListings.map((equipmentListing) => (
-                <Col key={equipmentListing.id} md={3}>
-                  <div className="row-container">
-                    <EquipmentCard
-                      equipmentListing={equipmentListing}
-                      onDeleteEquipment={onDeleteEquipment}
-                      onAddToCart={onAddToCart}
-                    />
-                  </div>
-                </Col>
-              ))}
-            </Row>
-          </Container>
-        </div>
+    <div className="equipment-page">
+      <div className="container">
+        <AddNewEquipment onAddEquipment={onAddEquipment} />
+        <Container fluid>
+          <Row>
+            {equipmentListings.map((equipmentListing) => (
+              <Col key={equipmentListing.id} md={3}>
+                <div className="row-container">
+                  <EquipmentCard
+                    equipmentListing={equipmentListing}
+                    onDeleteEquipment={onDeleteEquipment}
+                    onAddToCart={onAddToCart}
+                  />
+                </div>
+              </Col>
+            ))}
+          </Row>
+        </Container>
       </div>
-    </>
+    </div>
   )
 }
